+++ src/components/layouts/index.js
@@ -1,4 +1,3 @@
 export { default as Header } from './Header'
 export { default as Footer } from './Footer'
-export { default as Navigation } from './Navigation'
 export { default as MainLayout } from './MainLayout'
